% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comms.R
\docType{data}
\name{comms}
\alias{comms}
\title{dataset: metals commodities prices}
\format{
data frame
}
\usage{
comms
}
\value{
\code{tibble}
}
\description{
Adjusted metals commodity flat front-contract prices
}
\keyword{datasets}
