% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{dataset: metal producers prices}
\format{
data frame
}
\usage{
stocks
}
\value{
\code{tibble}
}
\description{
Adjusted metal producer stock prices
}
\keyword{datasets}
